fix unit tests rename files
