+++ openmetadata-ui/src/main/resources/ui/src/router/GlobalSettingRouter.tsx
@@ -79,7 +79,10 @@ const MsTeamsPage = withSuspenseFallback(
 );
 const ElasticSearchIndexPage = withSuspenseFallback(
   React.lazy(
-    () => import('../pages/ElasticSearchIndexPage/elastic-search-index.page')
+    () =>
+      import(
+        '../pages/ElasticSearchIndexPage/ElasticSearchReIndexPage.component'
+      )
   )
 );
